move keys to private/keys
this was confusing since there is also a keys.go in the root of the repo for the signing keypair.

right now these keys are used for private messages, so move the package there to make this clearer.
