+++ private/box2/spec_box_test.go
@@ -7,7 +7,7 @@ import (
 	"go.mindeco.de/ssb-refs/tfk"
 
 	"github.com/stretchr/testify/require"
-	"go.cryptoscope.co/ssb/keys"
+	"go.cryptoscope.co/ssb/private/keys"
 )
 
 type boxSpecTest struct {
