+++ private/keys/error.go
