+++ private/keys/ops_store_test.go
