+++ private/keys/store_test.go
