+++ private/manager.go
@@ -15,7 +15,7 @@ import (
 
 	"go.cryptoscope.co/ssb"
 	"go.cryptoscope.co/ssb/internal/extra25519"
-	"go.cryptoscope.co/ssb/keys"
+	"go.cryptoscope.co/ssb/private/keys"
 	"go.cryptoscope.co/ssb/private/box"
 	"go.cryptoscope.co/ssb/private/box2"
 	refs "go.mindeco.de/ssb-refs"
